+++ eeos/src/main/java/com/blackcompany/eeos/teamBuilding/application/service/CommandTeamBuildingTargetMemberService.java
@@ -6,8 +6,8 @@ import com.blackcompany.eeos.target.application.dto.TargetMember;
 import com.blackcompany.eeos.target.application.model.converter.TeamBuildingTargetEntityConverter;
 import com.blackcompany.eeos.target.application.service.CommandTargetMemberService;
 import com.blackcompany.eeos.target.application.service.SelectTargetMemberService;
-import com.blackcompany.eeos.target.persistence.TeamBuildingTargetEntity;
-import com.blackcompany.eeos.target.persistence.TeamBuildingTargetRepository;
+import com.blackcompany.eeos.target.persistence.teambuilding.TeamBuildingTargetEntity;
+import com.blackcompany.eeos.target.persistence.teambuilding.TeamBuildingTargetRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
